adds PostProcessor class.
this is used to implement a two pass rendering pipeline,
where in the first pass, the whole screen colors and depth
are rendered to a texture — which allows for postprocessing —
and in the second pass the postprocessed texture is rendered
to a screen filling quad.

currently a postprocess_invert_shader has been added and enabled.

postprocessing can be toggled using the F7 key.